+++ src/utils/dateUtils.js
@@ -1,0 +1,7 @@
+export const dateFormatter = new Intl.DateTimeFormat("en", {
+  month: "long",
+  weekday: "short",
+  day: "numeric",
+});
+
+export default { dateFormatter };
